Make acl build with current master

+++ caffe2/mobile/contrib/arm-compute/test/gl_operator_test.h
@@ -25,7 +25,7 @@ namespace caffe2 {
 // and unique at each value (for debugging purposes)
 template<typename T = float>
 void PopulateCPUBlob(Workspace *ws, bool random, std::string name,
-                     std::vector<int> dims, int val = 1, int dist_shift = 0, float variance = 1) {
+                     std::vector<int> dims, int val = 1, float dist_shift = 0, float variance = 1) {
   Blob *blob = ws->CreateBlob(name);
   auto* tensor = BlobGetMutableTensor(blob, CPU);
   tensor->Resize(dims);
